requirements.txt
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
click==8.0.1
distlib==0.3.2
dominate==2.6.0
elasticsearch==7.14.0
filelock==3.0.12
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.15.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pipenv==2021.5.29
platformdirs==2.3.0
python-dotenv==0.19.0
six==1.16.0
urllib3==1.26.6
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
